fixing some build warnings on gcc 4
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@785 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/util/impl/thread_posix.inl
@@ -206,8 +206,8 @@ public:
 	}	
 	const WaitResult wait(unsigned long iMilliSeconds)
 	{
-		const unsigned million =  1000000;
-		const unsigned trillion = 1000000000;
+		const long million =  1000000;
+		const long trillion = 1000000000;
 		
 		LASS_ENFORCE_CLIB_RC(pthread_mutex_lock(&mutex_));
